+++ NEWS.md
@@ -3,7 +3,7 @@
 ### Enhancements
 
 * Added option to choose which variables can be filtered in the filter panel by using the `filterable` attributes for the per-dataset lists in the `filter` argument of `init`. 
-* `teal_module` having `data` argument in its arguments will receive a list of reactive filter data with `"code"` and `"join_keys"` attributes.
+* `teal_module` having `data` argument in its arguments of type `tdata`. Modules receive this object (a list of reactive filtered datasets with `"code"` and `"join_keys"` attributes).
 * Updated the internals of `module_teal` to reflect changes in `teal.slice`.
 * `teal_module` having `filter_panel_api` argument in its arguments will receive a `FilterPanelAPI` object.